backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.8
psycopg2-binary==2.9.5
python-dotenv==1.0.0
bcrypt==4.0.1
python-multipart==0.0.6
networkx==3.1